Provide for redirects from project folders.
Per PEP-503, all URLs that return HTML5 must end with a /, and
we SHOULD redirect URLs without a trailing / to one with the /
appended. Suspect this means we will need the following objects
to exist:
* `project`
* `project/index.html`
where the former is a dummy object that has object-level
redirection configured on it to redirect to `/project/` and the
latter is what will actually be served from `/project/`.